move LocalHeaderBuffer to spec
