+++ src/spec.rs
@@ -592,3 +592,49 @@ impl Zip64CentralDirectoryEnd {
         Ok(())
     }
 }
+
+#[repr(packed)]
+pub struct LocalHeaderBuffer {
+    // local file header signature
+    pub magic: u32,
+    // version needed to extract
+    pub version_needed_to_extract: u16,
+    // general purpose bit flag
+    pub flag: u16,
+    // Compression method
+    pub compression_method: u16,
+    // last mod file time and last mod file date
+    pub last_modified_time_timepart: u16,
+    pub last_modified_time_datepart: u16,
+    // crc-32
+    pub crc32: u32,
+    // compressed size and uncompressed size
+    pub compressed_size: u32,
+    pub uncompressed_size: u32,
+    // file name length
+    pub file_name_length: u16,
+    // extra field length
+    pub extra_field_length: u16,
+}
+
+impl LocalHeaderBuffer {
+    #[inline]
+    pub fn extract(mut info: [u8; mem::size_of::<Self>()]) -> Self {
+        let start: *mut u8 = info.as_mut_ptr();
+
+        LittleEndian::from_slice_u16(unsafe { slice::from_raw_parts_mut(start as *mut u16, 15) });
+
+        unsafe { mem::transmute(info) }
+    }
+
+    #[inline]
+    pub fn writable_block(self) -> [u8; mem::size_of::<Self>()] {
+        let mut buf: [u8; mem::size_of::<Self>()] = unsafe { mem::transmute(self) };
+
+        LittleEndian::from_slice_u16(unsafe {
+            slice::from_raw_parts_mut(buf.as_mut_ptr() as *mut u16, 15)
+        });
+
+        buf
+    }
+}
